fixed offset for cut front sector

+++ diplom/GLSL/mesh.fs
@@ -1,7 +1,7 @@
 #include GLSL/lighting.fs
 
 uniform int sector;
-uniform int cutX, cutY, cutZ;
+uniform float cutX, cutY, cutZ;
 uniform int dimX, dimY, dimZ;
 
 uniform bool showFS;
